Add content access details to the `/user-attributes/me/membership` endpoint
This data is going under `/membership` rather than `/features` because
it directly discloses that the user is a paying customer, which we're going
to consider a piece of personal information.

Note also that we're going to want Staff to be able to see 'paid' member
content, even though they don't pay.

https://trello.com/c/KhTyeT4p/193-restricted-content-change-user-me-endpoint-to-member-data-api

https://members-data-api.theguardian.com/user-attributes/me/membership