Get accurate Membership status from Zuora on webhook update
We've had a problem since cross currency free-to-paid upgrades was
merged:

https://github.com/guardian/membership-frontend/pull/1124
https://github.com/guardian/membership-frontend/pull/1128

https://trello.com/c/qtBVUIN1/497-salesforce-and-mem-data-api-out-of-sync

...now that a new Zuora account is made on free-to-paid upgrade, we're
seeing the `Membership_Tier__c` field that comes through from the webhook
give the _old_ value rather than the new one. So now we're choosing to
get the info straight from the horse's mouth: Zuora.

Although the webhook passes the SF Contact Id, it doesn't pass the SF
Account Id, which we need for linking to Zuora (`crmId`) - so we have to
make the annoying `contactRepo.get(attrs.userId)` call to get it.