Parse Guardian Identity cookies and use them to perform lookup of stats
This means that the page view stats returned by the Friendly Tailor
/tailor/me endpoint will be for the logged-in user, across all their
devices, if the user is logged in. Otherwise the stats are based on the
Ophan browser cookie as before.

The response will include the user id if it found one, to make it clearer
where the stats are coming from.