HackDay: Allow users to verify themselves by donating to charity
When a user wants to post to a contentious article, they can verify
themselves by first donating to a nominated charity.

See JustGiving's 'Simple Donation Integration': https://developer.justgiving.com/apidocs/sdi

For instance, for this charity:

https://www.justgiving.com/fundraising/roberto-tyley

...we would construct a link like this, and offer it to the user as a
verification method:

http://www.justgiving.com/roberto-tyley/4w350m3/donate?amount=2.00&exitUrl=https%3A%2F%2Fmembers-data-api.theguardian.com%2Fverify%2Fjustgiving%3FreturnUrl%3Dhttp%253A%252F%252Fwww.theguardian.com%252Fpolitics%252F2016%252Fjun%252F10%252Fno-single-market-access-for-uk-after-brexit-wolfgang-schauble-says%26jgDonationId%3DJUSTGIVING-DONATION-ID&currency=GBP&reference=ghackday&defaultMessage=test%20message&utm_source=sdisponsor&utm_medium=buttons&utm_campaign=buttontype

JustGiving lets us specify an exit url that the user is redirected to
after the donation occurs. In our case, the exit url is this
`/verify/justgiving` endpoint on the members-data API:

https://members-data-api.theguardian.com/verify/justgiving?returnUrl=http%3A%2F%2Fwww.theguardian.com%2Fpolitics%2F2016%2Fjun%2F10%2Fno-single-market-access-for-uk-after-brexit-wolfgang-schauble-says

...note that this also accepts a return url(!). Once the donation has been
recorded to our database, we redirect to the page the user was originally
intending to comment upon. On that page, a call to the new `/reputation`
endpoint can be made, to see if the user now has enough reputation to
proceed.