Add Guardian Identity headers to HTTP response for Fastly logging
Antonio Favata (iOS developer on the Mobile team) had seen a recent spike
in new errors coming from the members-data-api - a total of 17K errors
starting Feb 1st, with a couple of spikes on Feb 2nd (3pm and 10pm).
This is in addition to the 'regular background noise' of empty-response
errors that the iOS app says it gets from the members-data-api.

After trying to investigate this, I feel that better logging on the
members-data-api is required to understand what's going on : specifically,
we need to turn on Fastly logging (it's not currently turned on for the
members-data-api), and we need to add Guardian Identity information to the
response headers so that it can be recorded in the Fastly logs. At the very
least, this should allow us to look at a specific example of an error and
say - "this is how Fastly says we responded to that request".

Unfortunately, although we normally just add the relevant headers with
`AddGuIdentityHeaders` filter, these response headers are thrown away
for `onBadRequest`, `onHandlerNotFound` and `onError` (eg a HTTP 500
internal server error response). Because we're particularly concerned
about these, I've added the Identity response headers into the error
handlers in `Global` as well.


https://trello.com/c/i3BzCE6i/320-add-fastly-logging-to-members-data-api-to-help-diagnose-errors