Updated the Attributes and Features models to contain a grouping of cards rather than just one, as an Attributes record now relates to multiple Zuora subscriptions/billing accounts.
I've also used this opportunity to tweak the /features endpoint to list which product an expired card is for and what the expiry date was. This means we no longer output non-expired card expiry dates via that endpoint (a tiny bit safer).

There are a few TODOs dotted around which explain how we'll merge the Salesforce hook with the /update-attributes endpoint, so that there's only one code path updating Dynamo.