WIP: Remove redundant zuora call
When finding existing payment methods, the code currently calls zuora
twice for account details: the first time to turn a contact ID into an
account ID (throwing away the rest of the account details in the
process), and the second time to get the account details again.

This change surfaces the account details from the first call within
ExistingPaymentOptionsController, and then removes the second call.

A couple of other places in the code also turn a contact ID into an
account ID in the same way: I haven’t yet checked whether they can avoid
a later call. For at least one of them, the later call is to zuora’s
rest API, which seems to return slightly more info – if any of that info
is used this would mean the call can’t be avoided, though perhaps the
soap call could be replaced by an up-front rest call.