Add SupporterPlus student rate plans to attributes mapper
Most systems use the benefits API to know which benefits to give to a
user based on their subscription(s). However the iOS live app still uses
members-data-api, which before this change didn't know to give benefits
to users with the SupporterPlus OneYearStudent rate plan. This meant
that users who have taken out the student offer were not getting the
benefits of their subscription in the iOS app. On the web and in the
Android app things will have worked fine as they use the
user-benefits-api.